Revert "Bump JOGL"
This reverts commit e1a9deec9b60f51612adf171df3118ea0363082b.

# Conflicts:
#	src/haven/HavenPanel.java

+++ src/haven/GLShader.java
@@ -28,7 +28,7 @@ package haven;
 
 import java.io.*;
 import java.util.*;
-import com.jogamp.opengl.*;
+import javax.media.opengl.*;
 
 public abstract class GLShader implements java.io.Serializable {
     public final String source, header;
@@ -50,7 +50,7 @@ public abstract class GLShader implements java.io.Serializable {
         }
 
         public void create(GL2 gl) {
-            id = (int)gl.glCreateShaderObjectARB(type); // FIXME: make long
+            id = gl.glCreateShaderObjectARB(type);
             GOut.checkerr(gl);
         }
